+++ internal/models/accrual.go
@@ -23,7 +23,7 @@ type Good struct {
 
 // OrderAccrualResp - модель ответа на запрос.
 type OrderAccrualResp struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual"`
 }
